+++ CDS_CarrotJob/CDS_CarrotJob/Data/Model/Home/JobModel.swift
@@ -1,0 +1,16 @@
+//
+//  job.swift
+//  CDS_CarrotJob
+//
+//  Created by 김동현 on 2023/05/16.
+//
+
+import UIKit
+
+struct JobLocalModel {
+    let jobImage: String
+    let mainInfo: String
+    let subInfo: String
+    let moneyInfo: Int
+    let postId: Int
+}
