+++ CDS_CarrotJob/CDS_CarrotJob/Data/Model/Home/PartTimeJobModel.swift
@@ -20,20 +20,6 @@ struct PartTimeJobModel {
     let jobTime: String
 }
 
-//extension PartTimeServerModel {
-//    
-//    static func fetchPartTimeJobServerData() -> [PartTimeServerModel] {
-//        return [
-//            PartTimeServerModel(jobName: "데이즈호텔 프론트", jobImage: "https://pds.joongang.co.kr/news/component/htmlphoto_mmdata/202203/24/6cff806a-82ef-4a54-a7ba-acd0ced51ba6.jpg", jobHourlyWage: 10000),
-//            PartTimeServerModel(jobName: "데이즈호텔 프론트", jobImage: "https://pds.joongang.co.kr/news/component/htmlphoto_mmdata/202203/24/6cff806a-82ef-4a54-a7ba-acd0ced51ba6.jpg", jobHourlyWage: 10000),
-//            PartTimeServerModel(jobName: "데이즈호텔 프론트", jobImage: "https://pds.joongang.co.kr/news/component/htmlphoto_mmdata/202203/24/6cff806a-82ef-4a54-a7ba-acd0ced51ba6.jpg", jobHourlyWage: 10000),
-//            PartTimeServerModel(jobName: "데이즈호텔 프론트", jobImage: "https://pds.joongang.co.kr/news/component/htmlphoto_mmdata/202203/24/6cff806a-82ef-4a54-a7ba-acd0ced51ba6.jpg", jobHourlyWage: 10000),
-//            PartTimeServerModel(jobName: "데이즈호텔 프론트", jobImage: "https://pds.joongang.co.kr/news/component/htmlphoto_mmdata/202203/24/6cff806a-82ef-4a54-a7ba-acd0ced51ba6.jpg", jobHourlyWage: 10000),
-//            PartTimeServerModel(jobName: "데이즈호텔 프론트", jobImage: "https://pds.joongang.co.kr/news/component/htmlphoto_mmdata/202203/24/6cff806a-82ef-4a54-a7ba-acd0ced51ba6.jpg", jobHourlyWage: 10000)
-//        ]
-//    }
-//}
-
 extension PartTimeJobModel {
     
     static func fetchPartTimeJobdummyData() -> [PartTimeJobModel] {
