+++ CDS_CarrotJob/CDS_CarrotJob/Presentation/Apply/View/ApplyNavigation.swift
@@ -60,9 +60,5 @@ extension ApplyNavigaitionBarView {
             $0.leading.trailing.equalToSuperview()
             $0.height.equalTo(1)
         }
-        
-        
     }
-    
 }
-
