--- CDS_CarrotJob/CDS_CarrotJob/Presentation/Home/View/home.swift
@@ -1,8 +1,0 @@
-//
-//  Home.swift
-//  CDS_CarrotJob
-//
-//  Created by KJ on 2023/05/16.
-//
-
-import Foundation
