--- CDS_CarrotJob/CDS_CarrotJob/Presentation/Review/View/Review.swift
@@ -1,8 +1,0 @@
-//
-//  Review.swift
-//  CDS_CarrotJob
-//
-//  Created by KJ on 2023/05/16.
-//
-
-import Foundation
